refactor(pb): move createProjectSchema
